backend/scripts/requirements.txt
# Use with python3.7.9 through python installer not pyenv
# cryptography will not install, must do following:
# brew install openssl@3 rust
# env OPENSSL_DIR="$(brew --prefix openssl@3)" pip install cryptography

# add_papers.py reqs
python-dotenv==0.21.1
elasticsearch==8.14.0
sentence-transformers==2.2.2
urllib3==2.0.7
feedparser==6.0.11

# lbnlp reqs
#grpcio==1.51 # pain in the ass to install as dep for tensorflow, must be this version
tensorflow==1.15.0
# pip install https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.5.0-py3-none-any.whl
numpy==1.18.4
gensim==3.7.1
monty==3.0.2
ChemDataExtractor==1.3.0
dill==0.3.0
Unidecode==1.1.1
pymatgen==2019.9.8
PubChemPy==1.0.4
tqdm==4.66.5
#mypy==1.4.1
